refactor: comprehensive rename Static→Derive across codebase
