+++ packages/i18n/src/index.ts
@@ -1,3 +1,3 @@
 export * from './translation-functions/msg';
-export * from './translation-functions/static';
+export * from './translation-functions/derive';
 export * from './translation-functions/fallbacks';
