+++ packages/react-core-linter/src/utils/constants.ts
@@ -16,8 +16,9 @@ export const VAR_COMPONENT_NAME = 'Var';
 export const NUM_COMPONENT_NAME = 'Num';
 export const CURRENCY_COMPONENT_NAME = 'Currency';
 export const DATE_TIME_COMPONENT_NAME = 'DateTime';
-export const STATIC_COMPONENT_NAME = 'Static';
 export const DERIVE_COMPONENT_NAME = 'Derive';
+/** @deprecated Use `DERIVE_COMPONENT_NAME` instead. */
+export const STATIC_COMPONENT_NAME = 'Static';
 export const BRANCH_COMPONENT_NAME = 'Branch';
 export const PLURAL_COMPONENT_NAME = 'Plural';
 
@@ -29,8 +30,11 @@ export const GET_TRANSLATIONS_FUNCTION_NAME = 'getTranslations';
 export const USE_MESSAGES_FUNCTION_NAME = 'useMessages';
 export const GET_MESSAGES_FUNCTION_NAME = 'getMessages';
 export const MSG_FUNCTION_NAME = 'msg';
+export const DERIVE_FUNCTION_NAME = 'derive';
+/** @deprecated Use `DERIVE_FUNCTION_NAME` instead. */
 export const DECLARE_STATIC_FUNCTION_NAME = 'declareStatic';
-export const DECLARE_DERIVE_FUNCTION_NAME = 'derive';
+/** @deprecated Use `DERIVE_FUNCTION_NAME` instead. */
+export const DECLARE_DERIVE_FUNCTION_NAME = DERIVE_FUNCTION_NAME;
 
 export const GT_COMPONENT_NAMES = [
   T_COMPONENT_NAME,
