+++ packages/react-core-linter/src/utils/isGTFunction.ts
@@ -132,7 +132,7 @@ export function isTComponent({
   });
 }
 
-export function isStaticComponent({
+export function isDeriveComponent({
   context,
   node,
   libs,
@@ -141,14 +141,19 @@ export function isStaticComponent({
     context,
     node,
     libs,
-    targetName: STATIC_COMPONENT_NAME,
+    targetName: DERIVE_COMPONENT_NAME,
   }) || isGTFunction({
     context,
     node,
     libs,
-    targetName: DERIVE_COMPONENT_NAME,
+    targetName: STATIC_COMPONENT_NAME,
   });
 }
+
+/**
+ * @deprecated Use `isDeriveComponent` instead.
+ */
+export const isStaticComponent = isDeriveComponent;
 
 export function isVariableComponent({
   context,
@@ -202,7 +207,7 @@ export function isBranchingComponent({
   });
 }
 
-export function isDeclareStaticFunction({
+export function isDeriveFunction({
   context,
   node,
   libs,
@@ -211,14 +216,19 @@ export function isDeclareStaticFunction({
     context,
     node,
     libs,
-    targetName: DECLARE_STATIC_FUNCTION_NAME,
+    targetName: DECLARE_DERIVE_FUNCTION_NAME,
   }) || isGTFunction({
     context,
     node,
     libs,
-    targetName: DECLARE_DERIVE_FUNCTION_NAME,
+    targetName: DECLARE_STATIC_FUNCTION_NAME,
   });
 }
+
+/**
+ * @deprecated Use `isDeriveFunction` instead.
+ */
+export const isDeclareStaticFunction = isDeriveFunction;
 
 export function isMsgFunction({
   context,
